Add `use` statement for DeviceExt
I tripped over this one while following along with the tutorial. I added a little link to extension traits as I'm aware the audience of this tutorial may not be familiar with them.

Please feel free to rewrite/throw out any of my copy as you see fit.  😄 

Thanks again for this phenomenal tutorial!