Update Surface tutorial State::new()
Adds an additional explanation that the `event_loop` needs to be updated after the window has been moved. While this is fixed in the input() section, including this mention in new() allows the program to compile during the new() and resize() section if the reader did not catch this themselves. 

Adds clarity following "changed State to own window" (09207e9) commit