Remove an inaccurate comment.
This comment is inaccurate. Drop order doesn't matter in this case,
because this struct only holds a reference. Dropping it does nothing,
the surface has its own reference to the window.

The actual window is owned by the scope outside of `State` and thus
references to it that are kept in the `State` are valid.

The lifetime of the `Surface` takes care of not allowing for dropping
the window while the surface exists.

I also removed a stray space at the end of a line.