Use `div_ceil` where appropriate. (#631)
Several places in the tutorial need to do a division that rounds up, for
chunking operations. They use a form like `n / k + (n % k > 0)`.

I'm used to seeing this kind of division done with `(n + (k - 1)) / k`.
This is probably preferred because it avoids the `%` operation, which is
typically expensive.

But Rust has a builtin for this: `div_ceil`. Even better.